Ability to specify both region and account ID of ECR registry
For our use case we store all of our ECR repositories in a single
account, and access these repos from others. Therefore we require the
need to specify the account which possess the repostories, along with
the region.

The region wasn't a necessary change however I have seen others
requesting this and it was an easy implementation alongside this work.