Amend default region variable
Reason for this is because when trying to grab the region via the means
previously implemented it was unsuccessful on a Buildkite agent running
on EC2.

As we have changed to using `get-login-password` due to the previous
implementation being deprecated, it requires the region as well as the
account id to be given as part of the docker login command.

Ideally I would leave the region out of these changes entirely.

If nothing is given both via the `region` flag or `AWS_DEFAULT_REGION`
env var, it will default to `eu-west-1`.